StudyHome: Move folder to new location
