+++ src/components/StudyHomeComponents/RecentStudies.js
